# Project/1.PondPlanning/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-based Village Pond Planning System",
    "description": "Terrain, catchment and runoff analysis for village pond siting. Analysis routes are asynchronous: they return 202 with a job id to poll.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/villages": {
      "get": {
        "tags": [
          "villages"
        ],
        "summary": "List villages",
        "description": "Return the registered villages, newest first.",
        "operationId": "list_villages_api_v1_villages_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive name search",
              "title": "Q"
            },
            "description": "Case-insensitive name search"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_VillageOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "villages"
        ],
        "summary": "Create Village",
        "description": "Register a village from a name and a GeoJSON boundary.\n\nThe UTM zone for every later computation is derived from this boundary's own\ncentroid at creation time \u2014 never configured, never hard-coded.",
        "operationId": "create_village_api_v1_villages_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VillageCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VillageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/villages/{village_id}": {
      "get": {
        "tags": [
          "villages"
        ],
        "summary": "Village detail",
        "description": "Return one village.",
        "operationId": "get_village_api_v1_villages__village_id__get",
        "parameters": [
          {
            "name": "village_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Village identifier",
              "title": "Village Id"
            },
            "description": "Village identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VillageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/villages/{village_id}/summary": {
      "get": {
        "tags": [
          "villages"
        ],
        "summary": "Get Village Summary",
        "description": "FR1: area, elevation range and mean slope for the selected village.",
        "operationId": "get_village_summary_api_v1_villages__village_id__summary_get",
        "parameters": [
          {
            "name": "village_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Village identifier",
              "title": "Village Id"
            },
            "description": "Village identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VillageSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/villages/{village_id}/imagery": {
      "get": {
        "tags": [
          "villages"
        ],
        "summary": "Get Village Imagery",
        "description": "FR1: the satellite basemap descriptor, clipped to this village.",
        "operationId": "get_village_imagery_api_v1_villages__village_id__imagery_get",
        "parameters": [
          {
            "name": "village_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Village identifier",
              "title": "Village Id"
            },
            "description": "Village identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageryLayer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/villages/{village_id}/available-land": {
      "get": {
        "tags": [
          "villages"
        ],
        "summary": "Get Available Land",
        "description": "FR3: land eligible for excavation, with the constraints that produced it.",
        "operationId": "get_available_land_api_v1_villages__village_id__available_land_get",
        "parameters": [
          {
            "name": "village_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Village identifier",
              "title": "Village Id"
            },
            "description": "Village identifier"
          },
          {
            "name": "max_slope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 45,
              "minimum": 0,
              "default": 5.0,
              "title": "Max Slope"
            }
          },
          {
            "name": "min_area_ha",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "exclusiveMinimum": 0,
              "default": 0.5,
              "title": "Min Area Ha"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailableLandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/villages/{village_id}/parcels:import": {
      "post": {
        "tags": [
          "villages"
        ],
        "summary": "Import Parcels",
        "description": "Import cadastral parcels from a zipped Shapefile or GeoJSON.\n\nAsynchronous because a district parcel set is large. Upload hardening\n(driver whitelist, zip-entry validation, size cap) lands with the engine in\nP4; the contract is fixed here.",
        "operationId": "import_parcels_api_v1_villages__village_id__parcels_import_post",
        "parameters": [
          {
            "name": "village_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Village identifier",
              "title": "Village Id"
            },
            "description": "Village identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_parcels_api_v1_villages__village_id__parcels_import_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AvailableLandResponse": {
        "properties": {
          "village_id": {
            "type": "string",
            "format": "uuid",
            "title": "Village Id"
          },
          "constraints_applied": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Constraints Applied"
          },
          "total_eligible_area": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "parcels": {
            "items": {
              "$ref": "#/components/schemas/LandParcel"
            },
            "type": "array",
            "title": "Parcels"
          },
          "geojson": {
            "$ref": "#/components/schemas/GeoJSONFeatureCollection"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ResultWarning"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "village_id",
          "constraints_applied",
          "total_eligible_area",
          "parcels",
          "geojson"
        ],
        "title": "AvailableLandResponse",
        "description": "FR3: eligible excavation land, with the constraint set that produced it."
      },
...
      "Body_import_parcels_api_v1_villages__village_id__parcels_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_parcels_api_v1_villages__village_id__parcels_import_post"
      },
...
      "ElevationSummary": {
        "properties": {
          "minimum": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "maximum": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "mean": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "relief": {
            "$ref": "#/components/schemas/QuantityOut",
            "description": "maximum - minimum"
          }
        },
        "type": "object",
        "required": [
          "minimum",
          "maximum",
          "mean",
          "relief"
        ],
        "title": "ElevationSummary",
        "description": "Elevation statistics over the village boundary."
      },
...
      "GeoJSONFeature": {
        "properties": {
          "type": {
            "type": "string",
            "const": "Feature",
            "title": "Type",
            "default": "Feature"
          },
          "geometry": {
            "additionalProperties": true,
            "type": "object",
            "title": "Geometry"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties"
          }
        },
        "type": "object",
        "required": [
          "geometry"
        ],
        "title": "GeoJSONFeature",
        "description": "A GeoJSON Feature, loosely typed.\n\nGeometry is not modelled field-by-field on purpose: the map client consumes\nit directly, and a hand-written GeoJSON type is a large surface that adds no\nsafety the client actually uses. Geometry validity is enforced in PostGIS and\nin the engines, where it matters."
      },
      "GeoJSONFeatureCollection": {
        "properties": {
          "type": {
            "type": "string",
            "const": "FeatureCollection",
            "title": "Type",
            "default": "FeatureCollection"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/GeoJSONFeature"
            },
            "type": "array",
            "title": "Features"
          },
          "crs": {
            "type": "string",
            "title": "Crs",
            "description": "Always EPSG:4326 on the wire",
            "default": "EPSG:4326"
          }
        },
        "type": "object",
        "required": [
          "features"
        ],
        "title": "GeoJSONFeatureCollection",
        "description": "A GeoJSON FeatureCollection, always in EPSG:4326 on the wire.\n\nEvery computation happens in the UTM zone derived from the data's own\ncentroid; 4326 is the interchange format only. ``crs`` is stated explicitly\nso no consumer has to assume."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImageryLayer": {
        "properties": {
          "layer_id": {
            "type": "string",
            "title": "Layer Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "tile_url_template": {
            "type": "string",
            "title": "Tile Url Template",
            "description": "XYZ template for the map client"
          },
          "attribution": {
            "type": "string",
            "title": "Attribution"
          },
          "captured": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured",
            "description": "Acquisition date or range, if known"
          },
          "max_zoom": {
            "type": "integer",
            "title": "Max Zoom"
          }
        },
        "type": "object",
        "required": [
          "layer_id",
          "provider",
          "tile_url_template",
          "attribution",
          "max_zoom"
        ],
        "title": "ImageryLayer",
        "description": "FR1: a satellite basemap clipped to the village boundary."
      },
      "JobAccepted": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "const": "queued",
            "title": "Status",
            "default": "queued"
          },
          "poll_url": {
            "type": "string",
            "title": "Poll Url",
            "description": "GET this until status is succeeded or failed"
          },
          "estimated_seconds": {
            "type": "integer",
            "title": "Estimated Seconds",
            "description": "Rough expectation, for the progress UI"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "poll_url",
          "estimated_seconds"
        ],
        "title": "JobAccepted",
        "description": "``202`` response for every long-running analysis.\n\nTerrain analysis takes tens of seconds. Rather than hold a connection open\nand hope no proxy times it out, every analysis route returns this immediately\nand the client polls ``poll_url``."
      },
...
      "LandParcel": {
        "properties": {
          "parcel_id": {
            "type": "string",
            "title": "Parcel Id"
          },
          "ownership_class": {
            "type": "string",
            "enum": [
              "government",
              "community",
              "private",
              "unknown"
            ],
            "title": "Ownership Class",
            "description": "Ownership *class*, never an owner's name \u2014 personal data under the DPDP Act 2023. See ADR 0012."
          },
          "area": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "mean_slope": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "lulc_class": {
            "type": "string",
            "title": "Lulc Class"
          },
          "excluded_by": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Excluded By",
            "description": "Constraints that rejected this parcel, if any"
          },
          "eligible": {
            "type": "boolean",
            "title": "Eligible"
          }
        },
        "type": "object",
        "required": [
          "parcel_id",
          "ownership_class",
          "area",
          "mean_slope",
          "lulc_class",
          "eligible"
        ],
        "title": "LandParcel",
        "description": "FR3: one polygon of land judged available for excavation."
      },
...
      "Page_VillageOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VillageOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[VillageOut]"
      },
...
      "QuantityOut": {
        "properties": {
          "value": {
            "type": "number",
            "title": "Value"
          },
          "unit": {
            "$ref": "#/components/schemas/Unit"
          },
          "uncertainty_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uncertainty Pct",
            "description": "Symmetric relative uncertainty in percent. null means exact by construction."
          },
          "low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low",
            "description": "Lower bound of the uncertainty band"
          },
          "high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "High",
            "description": "Upper bound of the uncertainty band"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method",
            "description": "How the value was derived"
          },
          "display": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display",
            "description": "Preformatted for direct rendering"
          }
        },
        "type": "object",
        "required": [
          "value",
          "unit"
        ],
        "title": "QuantityOut",
        "description": "A number with its unit, uncertainty band and provenance.\n\nThe wire counterpart of :class:`app.domain.units.Quantity`. ``display`` is\nincluded deliberately: it removes any chance of the frontend rendering\n\"18950\" because someone forgot to append the unit, and it is what appears in\nreport screenshots.",
        "example": {
          "display": "18,950.00 m\u00b3 (\u00b120 %)",
          "high": 22740.0,
          "low": 15160.0,
          "method": "EAV curve integrated over the filled depression",
          "uncertainty_pct": 20.0,
          "unit": "m3",
          "value": 18950.0
        }
      },
...
      "ResultWarning": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "caution",
              "critical"
            ],
            "title": "Severity",
            "default": "caution"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ResultWarning",
        "description": "A caveat attached to an otherwise successful result.\n\nWarnings are how this system avoids presenting a number more confidently than\nit deserves \u2014 \"the source DEM is 30 m, so relief below ~5 m is interpolated,\nnot measured\" belongs next to the storage estimate, not buried in a report."
      },
...
      "Unit": {
        "type": "string",
        "enum": [
          "m",
          "m2",
          "ha",
          "km2",
          "m3",
          "mm",
          "mm/yr",
          "deg",
          "%",
          "ratio",
          "count",
          "yr",
          "d",
          "INR"
        ],
        "title": "Unit",
        "description": "Units used anywhere in the system. SI, except where the domain differs.\n\nIndian minor-irrigation practice reports catchment in hectares and storage in\ncubic metres, so those are the reporting units even though the engine\ncomputes in square metres."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VillageCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          },
          "state_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "State Code"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "boundary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Boundary",
            "description": "GeoJSON MultiPolygon or Polygon, EPSG:4326"
          }
        },
        "type": "object",
        "required": [
          "name",
          "boundary"
        ],
        "title": "VillageCreate",
        "description": "Register a village by name and boundary."
      },
      "VillageOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "state_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Code"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "centroid": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Centroid",
            "description": "[lon, lat] in EPSG:4326"
          },
          "utm_epsg": {
            "type": "integer",
            "title": "Utm Epsg",
            "description": "Projected CRS used for every area and distance computation, derived at runtime from this village's own centroid \u2014 never configured, never hard-coded."
          },
          "area": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "state_code",
          "district",
          "centroid",
          "utm_epsg",
          "area",
          "created_at"
        ],
        "title": "VillageOut",
        "description": "A registered village."
      },
      "VillageSummary": {
        "properties": {
          "village": {
            "$ref": "#/components/schemas/VillageOut"
          },
          "elevation": {
            "$ref": "#/components/schemas/ElevationSummary"
          },
          "mean_slope": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "dem_source": {
            "type": "string",
            "title": "Dem Source",
            "description": "Provider and native resolution, for attribution"
          },
          "dem_vertical_accuracy": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ResultWarning"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "village",
          "elevation",
          "mean_slope",
          "dem_source",
          "dem_vertical_accuracy"
        ],
        "title": "VillageSummary",
        "description": "FR1: the headline card shown when a village is selected.\n\nEvery field is a :class:`QuantityOut`, so the panel cannot render a bare\nnumber even by accident."
      }
    }
  }
}